Adds racket and python implementations.
The racket implementation is just another version of the simulator. The Python version, on the other hand, is a dynamic program that computes exact probabilities quite quickly (where "exact" means "exact up to the problems of compounding floating-point error").

In python there's a recursion limit problem. For large numbers you may need to call sys.setrecursionlimit() and set it to a larger number than the default.